QuikQuantV2 Project Files
This updated version of QuikQuant has a few updates and generally works better.
Notable changes:
1. Calendar dropdown instead of making user manually enter dates in "YYYY-MM-DD" format.
2. Program saves API Key and when it has one it does not ask for another one but instead provides a button that says "Change API Key" that the user may use if they need to change it.
3. Changed "Stocks" button to more intuitive "Show Calculations and Graphs"

Biggest Issue:

For some reason, if there are more than one stocks displayed on infoPrintScreen, the "Show Calculations and Graphs" button simply doesn't work. Quick fixed it so that it doesn't throw an error, so now it just doesn't do anything when clicked if multiple stocks are present.